Refactor shared module components in terms of directory hierarchy. This way we follow the same convention inside the shared package as we've done in the other modules.
