+++ src/main/tv/codely/scala_http_api/module/shared/bus/domain/MessagePublisher.scala
@@ -1,4 +1,4 @@
-package tv.codely.scala_http_api.module.shared.domain
+package tv.codely.scala_http_api.module.shared.bus.domain
 
 trait MessagePublisher {
   def publish[T <: Message](message: T): Unit
