+++ src/main/tv/codely/scala_http_api/module/shared/dependency_injection/infrastructure/SharedModuleDependencyContainer.scala
@@ -1,19 +1,19 @@
-package tv.codely.scala_http_api.module.shared.infrastructure.dependency_injection
+package tv.codely.scala_http_api.module.shared.dependency_injection.infrastructure
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import tv.codely.scala_http_api.module.shared.domain.{Logger, MessagePublisher}
-import tv.codely.scala_http_api.module.shared.infrastructure.config.{DbConfig, MessageBrokerConfig}
-import tv.codely.scala_http_api.module.shared.infrastructure.logger.scala_logging.ScalaLoggingLogger
-import tv.codely.scala_http_api.module.shared.infrastructure.message_broker.rabbitmq.{RabbitMqChannelFactory, RabbitMqMessagePublisher}
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
+import tv.codely.scala_http_api.module.shared.bus.domain.MessagePublisher
+import tv.codely.scala_http_api.module.shared.bus.infrastructure.rabbit_mq.{RabbitMqChannelFactory, RabbitMqConfig, RabbitMqMessagePublisher}
+import tv.codely.scala_http_api.module.shared.logger.domain.Logger
+import tv.codely.scala_http_api.module.shared.logger.infrastructure.scala_logging.ScalaLoggingLogger
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.{DoobieDbConnection, JdbcConfig}
 
 import scala.concurrent.ExecutionContext
 
 final class SharedModuleDependencyContainer(
     actorSystemName: String,
-    dbConfig: DbConfig,
-    publisherConfig: MessageBrokerConfig
+    dbConfig: JdbcConfig,
+    publisherConfig: RabbitMqConfig
 ) {
   implicit val actorSystem: ActorSystem  = ActorSystem(actorSystemName)
   val materializer: ActorMaterializer    = ActorMaterializer()
