+++ src/main/tv/codely/scala_http_api/module/shared/persistence/infrastructure/doobie/JdbcConfig.scala
@@ -1,0 +1,14 @@
+package tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie
+
+import com.typesafe.config.Config
+
+object JdbcConfig {
+  def apply(config: Config): JdbcConfig = JdbcConfig(
+    driver = config.getString("driver"),
+    url = config.getString("url"),
+    user = config.getString("user"),
+    password = config.getString("password")
+  )
+}
+
+final case class JdbcConfig(driver: String, url: String, user: String, password: String)
