+++ src/main/tv/codely/scala_http_api/module/shared/user/infrastructure/marshaller/UserIdJsonFormatMarshaller.scala
@@ -3,7 +3,7 @@ package tv.codely.scala_http_api.module.shared.user.infrastructure.marshaller
 import java.util.UUID
 
 import spray.json.{DefaultJsonProtocol, JsValue, JsonFormat, _}
-import tv.codely.scala_http_api.module.shared.infrastructure.marshaller.UuidJsonFormatMarshaller._
+import tv.codely.scala_http_api.module.shared.marshaller.infrastructure.UuidJsonFormatMarshaller._
 import tv.codely.scala_http_api.module.shared.user.domain.UserId
 
 object UserIdJsonFormatMarshaller extends DefaultJsonProtocol {
