+++ src/main/tv/codely/scala_http_api/module/user/infrastructure/dependency_injection/UserModuleDependencyContainer.scala
@@ -1,7 +1,7 @@
 package tv.codely.scala_http_api.module.user.infrastructure.dependency_injection
 
-import tv.codely.scala_http_api.module.shared.domain.MessagePublisher
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
+import tv.codely.scala_http_api.module.shared.bus.domain.MessagePublisher
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.DoobieDbConnection
 import tv.codely.scala_http_api.module.user.application.register.UserRegistrar
 import tv.codely.scala_http_api.module.user.application.search.UsersSearcher
 import tv.codely.scala_http_api.module.user.domain.UserRepository
