+++ src/main/tv/codely/scala_http_api/module/user/infrastructure/repository/DoobieMySqlUserRepository.scala
@@ -2,8 +2,8 @@ package tv.codely.scala_http_api.module.user.infrastructure.repository
 
 import doobie.implicits._
 import tv.codely.scala_http_api.module.user.domain.{User, UserRepository}
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.TypesConversions._
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.TypesConversions._
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.DoobieDbConnection
 
 import scala.concurrent.{ExecutionContext, Future}
