+++ src/main/tv/codely/scala_http_api/module/video/infrastructure/dependency_injection/VideoModuleDependencyContainer.scala
@@ -1,7 +1,7 @@
 package tv.codely.scala_http_api.module.video.infrastructure.dependency_injection
 
-import tv.codely.scala_http_api.module.shared.domain.MessagePublisher
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
+import tv.codely.scala_http_api.module.shared.bus.domain.MessagePublisher
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.DoobieDbConnection
 import tv.codely.scala_http_api.module.video.application.create.VideoCreator
 import tv.codely.scala_http_api.module.video.application.search.VideosSearcher
 import tv.codely.scala_http_api.module.video.domain.VideoRepository
