+++ src/main/tv/codely/scala_http_api/module/video/infrastructure/repository/DoobieMySqlVideoRepository.scala
@@ -1,9 +1,9 @@
 package tv.codely.scala_http_api.module.video.infrastructure.repository
 
 import doobie.implicits._
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
 import tv.codely.scala_http_api.module.video.domain.{Video, VideoRepository}
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.TypesConversions._
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.TypesConversions._
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.DoobieDbConnection
 
 import scala.concurrent.{ExecutionContext, Future}
