+++ src/test/tv/codely/scala_http_api/entry_point/AcceptanceSpec.scala
@@ -6,9 +6,9 @@ import akka.util.ByteString
 import com.typesafe.config.ConfigFactory
 import org.scalatest.{Matchers, WordSpec}
 import org.scalatest.concurrent.ScalaFutures
-import tv.codely.scala_http_api.module.shared.infrastructure.config.{DbConfig, MessageBrokerConfig}
-import tv.codely.scala_http_api.module.shared.infrastructure.dependency_injection.SharedModuleDependencyContainer
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
+import tv.codely.scala_http_api.module.shared.bus.infrastructure.rabbit_mq.RabbitMqConfig
+import tv.codely.scala_http_api.module.shared.dependency_injection.infrastructure.SharedModuleDependencyContainer
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.{DoobieDbConnection, JdbcConfig}
 import tv.codely.scala_http_api.module.user.infrastructure.dependency_injection.UserModuleDependencyContainer
 import tv.codely.scala_http_api.module.video.infrastructure.dependency_injection.VideoModuleDependencyContainer
 
@@ -20,8 +20,8 @@ protected[entry_point] abstract class AcceptanceSpec
   private val actorSystemName = "scala-http-api-acceptance-test"
 
   private val appConfig       = ConfigFactory.load("application")
-  private val dbConfig        = DbConfig(appConfig.getConfig("database"))
-  private val publisherConfig = MessageBrokerConfig(appConfig.getConfig("message-publisher"))
+  private val dbConfig        = JdbcConfig(appConfig.getConfig("database"))
+  private val publisherConfig = RabbitMqConfig(appConfig.getConfig("message-publisher"))
 
   private val sharedDependencies = new SharedModuleDependencyContainer(actorSystemName, dbConfig, publisherConfig)
