+++ src/test/tv/codely/scala_http_api/module/IntegrationTestCase.scala
@@ -1,11 +1,11 @@
 package tv.codely.scala_http_api.module
 
 import com.typesafe.config.ConfigFactory
-import tv.codely.scala_http_api.module.shared.domain.{Logger, MessagePublisher}
-import tv.codely.scala_http_api.module.shared.infrastructure.config.{DbConfig, MessageBrokerConfig}
-import tv.codely.scala_http_api.module.shared.infrastructure.dependency_injection.SharedModuleDependencyContainer
-import tv.codely.scala_http_api.module.shared.infrastructure.message_broker.rabbitmq.RabbitMqChannelFactory
-import tv.codely.scala_http_api.module.shared.infrastructure.persistence.doobie.DoobieDbConnection
+import tv.codely.scala_http_api.module.shared.bus.domain.MessagePublisher
+import tv.codely.scala_http_api.module.shared.bus.infrastructure.rabbit_mq.{RabbitMqChannelFactory, RabbitMqConfig}
+import tv.codely.scala_http_api.module.shared.dependency_injection.infrastructure.SharedModuleDependencyContainer
+import tv.codely.scala_http_api.module.shared.logger.domain.Logger
+import tv.codely.scala_http_api.module.shared.persistence.infrastructure.doobie.{DoobieDbConnection, JdbcConfig}
 
 import scala.concurrent.ExecutionContext
 
@@ -13,8 +13,8 @@ protected[scala_http_api] trait IntegrationTestCase extends UnitTestCase {
   private val actorSystemName = "scala-http-api-integration-test"
 
   private val appConfig       = ConfigFactory.load("application")
-  private val dbConfig        = DbConfig(appConfig.getConfig("database"))
-  private val publisherConfig = MessageBrokerConfig(appConfig.getConfig("message-publisher"))
+  private val dbConfig        = JdbcConfig(appConfig.getConfig("database"))
+  private val publisherConfig = RabbitMqConfig(appConfig.getConfig("message-publisher"))
 
   private val sharedDependencies = new SharedModuleDependencyContainer(actorSystemName, dbConfig, publisherConfig)
