+++ src/test/tv/codely/scala_http_api/module/shared/infrastructure/MessagePublisherMock.scala
@@ -2,7 +2,7 @@ package tv.codely.scala_http_api.module.shared.infrastructure
 
 import org.scalamock.scalatest.MockFactory
 import tv.codely.scala_http_api.module.UnitTestCase
-import tv.codely.scala_http_api.module.shared.domain.{Message, MessagePublisher}
+import tv.codely.scala_http_api.module.shared.bus.domain.{Message, MessagePublisher}
 
 protected[module] trait MessagePublisherMock extends MockFactory {
   this: UnitTestCase => // Make mandatory to also extend UnitTestCase in order to avoid using mocks in any other kind of test.
